requirements.txt
Django==1.7
django-reversion==1.8.4
wsgiref==0.1.2
